Fixed local camera not displaying
There was a glitch that caused your local camera to stop displaying video if you turned it off and then on.
This was fixed by modifying the useEffect hook to be activated when participant.videoEnabled is changed, which will pause the current video "v" when the camera is disabled, and then play it again when its enabled.